requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.112.0
frozenlist==1.4.1
general_operator==0.4.6
general_util==0.0.17
grpcio==1.65.4
grpcio-tools==1.65.4
h11==0.14.0
idna==3.7
influxdb-client==1.45.0
multidict==6.0.5
protobuf==5.27.3
psycopg2==2.9.10
pydantic==2.8.2
pydantic_core==2.20.1
PyMySQL==1.1.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
reactivex==4.0.4
redis==5.0.8
requests==2.32.3
setuptools==72.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.5
yarl==1.9.4
